+++ src/Step.js
@@ -1,12 +1,10 @@
 const Base = require('./Base')
-const Shell = require('./Shell')
 
 class Step extends Base {
   constructor (name, runtime) {
     super()
     this.name = name
     this.runtime = runtime
-    this.shell = new Shell(runtime)
   }
 }