+++ src/steps/Deps.js
@@ -1,5 +1,6 @@
 const Step = require('../Step')
 const utils = require('../Utils')
+
 // const debug = require('depurar')('frey')
 class Deps extends Step {
   main (cargo, cb) {